fix: scheduleService 수정 - timeFormat
