+++ src/controllers/scheduleController.ts
@@ -4,7 +4,6 @@ const { validationResult } = require('express-validator');
 import { userService, groupService, scheduleService } from "../services";
 import mongoose from "mongoose";
 import Schedule from "../models/Schedule";
-const setTimeFormat = require("./timeController");
 
 /**
  *  @route POST /schedule/:groupId
@@ -82,8 +81,8 @@ const getDailySchedule = async (req: Request, res: Response) => {
         const date = new Date(req.params.date);
         const day = Math.ceil((date.getTime() - group.startDate.getTime())/86400000)+1;
 
-        const schedules = await scheduleService.findSchedulesByDate(date, groupId);   // 해당 날짜 스케쥴 찾기
-        
+        const schedules = await scheduleService.
+        findSchedulesByDate(date, groupId);   // 해당 날짜 스케쥴 찾기
         const data = { day, date, schedules }; 
         
         return res.status(sc.OK).json({
